fix: permission of created .config dir in labwc
When the $HOMEDIR/.config/labwc folder is created it also creates the
$HOMEDIR/.config folder if it doesn't exist. As raspi-config is run with
sudo this makes the root user own the $HOMEDIR/.config directory.
Afterwards, only the $HOMEDIR/.config/labwc directory owner is changed
to $USER.

Add a function that ensures that the $HOMEDIR/.config directory exists
and, if it doesn't, create it with the correct permissions.

Fixes: https://github.com/RPi-Distro/raspi-config/issues/267
Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>